Use smaller type for max_frames
Not really to save memory, but semantically it's closer to what we want.

+++ ext/datadog_profiling_native_extension/collectors_stack.c
@@ -15,7 +15,7 @@ static VALUE missing_string = Qnil;
 
 // Used as scratch space during sampling
 struct sampling_buffer {
-  unsigned int max_frames;
+  uint16_t max_frames;
   ddog_prof_Location *locations;
   VALUE *stack_buffer;
   int *lines_buffer;
